+++ jar/src/main/java/pl/bookingsystem/db/dao/ClientDAO.java
@@ -5,7 +5,7 @@ import pl.bookingsystem.db.entity.Client;
 import java.io.PrintStream;
 import java.util.List;
 
-public interface ClientDAO extends GenericDAO<Client, Integer> {
+public interface ClientDAO extends GenericDAO<Client, Long> {
 
     public Client findByClientName(String name, String surname);
